update: Rename the proptypes folder to try to make import with alias work
