+++ package.json
@@ -1,5 +1,12 @@
 {
-  "name": "myreactpixelmosaic",
+  "name": "repixelact",
+  "imports": {
+    "@*": [
+      "./src/*",
+      "./src/*.ts",
+      "./src/*.tsx"
+    ]
+  },
   "private": true,
   "version": "0.0.0",
   "type": "module",
@@ -30,7 +37,7 @@
   "devDependencies": {
     "@types/node": "^20.11.25",
     "@types/parse-package-name": "^0.1.3",
-    "@types/react": "^18.2.64",
+    "@types/react": "^18.2.66",
     "@types/react-dom": "^18.2.19",
     "@typescript-eslint/eslint-plugin": "^7.1.1",
     "@typescript-eslint/parser": "^7.1.1",
