--- src/interfaces/font-loading.d.ts
@@ -1,7 +1,0 @@
-export interface FontFaceSet {
-  add(font: FontFace): void;
-}
-
-export interface Document {
-  fonts: FontFaceSet;
-}
