+++ tsconfig.json
@@ -2,12 +2,34 @@
   "compilerOptions": {
     "strict": true,
     "allowJs": false,
-    "target": "ESNext",
+    "target": "esnext",
+    "baseUrl": ".",
+    "paths": {
+      "*": ["src/*"],
+      "@assets/*": ["src/assets/*"],
+      "@common/*": ["src/common/*"],
+      "@components/*": ["src/components/*"],
+      "@config/*": ["src/config/*"],
+      "@constants/*": ["src/constants/*"],
+      "@containers/*": ["src/containers/*"],
+      "@contexts/*": ["src/contexts/*"],
+      "@env/*": ["src/env/*"],
+      "@features/*": ["src/features/*"],
+      "@reducers/*": ["src/features/reducers/*"],
+      "@fonts/*": ["src/fonts/*"],
+      "@helpers/*": ["src/helpers/*"],
+      "@hooks/*": ["src/hooks/*"],
+      "@interfaces/*": ["src/interfaces/*"],
+      "@layouts/*": ["src/layouts/*"],
+      "@services/*": ["src/services/*"],
+      "@less/*": ["src/styles/less/*"],
+      "@utils/*": ["src/utils/*"]
+    },
     "module": "ESNext",
     "moduleResolution": "Node",
     "skipLibCheck": true,
     "esModuleInterop": true,
-    "typeRoots": ["./src/@types", "./node_modules/@types"],
+    "typeRoots": ["./src/@types", "./node_modules/@types", "./src/@interfaces"],
     "outDir": "./build",
     "allowSyntheticDefaultImports": true,
     "forceConsistentCasingInFileNames": true,
@@ -16,14 +38,16 @@
     "isolatedModules": true,
     "noEmit": true,
     "types": ["node"],
-    "baseUrl": "./",
-    "paths": {
-      "~/*": ["src/*"]
-    },
     "lib": ["DOM", "DOM.Iterable", "ESNext"],
     "jsx": "react-jsx"
   },
-  "include": ["./src/**/*"],
+  "include": [
+    "prop-types.d.ts",
+    "font-loading.d.ts",
+    "index.d.ts",
+    "**/*.ts",
+    "**/*.tsx"
+  ],
   "exclude": ["node_modules"],
   "references": [{ "path": "./tsconfig.node.json" }]
 }
