+++ tsconfig.node.json
@@ -1,6 +1,8 @@
 {
   "compilerOptions": {
     "jsx": "react-jsx",
+    "strict": true,
+    "target": "esnext",
     "composite": true,
     "module": "ESNext",
     "moduleResolution": "Node"
