+++ vite.config.ts
@@ -7,7 +7,25 @@ import path from "path";
 export default defineConfig({
   resolve: {
     alias: {
-      "@": path.resolve(__dirname, "src")
+      "@": path.resolve(__dirname, "./src"),
+      "@assets": path.resolve(__dirname, "./src/assets"),
+      "@common": path.resolve(__dirname, "./src/common"),
+      "@components": path.resolve(__dirname, "./src/components"),
+      "@config": path.resolve(__dirname, "./src/config"),
+      "@constants": path.resolve(__dirname, "./src/constants"),
+      "@containers": path.resolve(__dirname, "./src/containers"),
+      "@contexts": path.resolve(__dirname, "./src/contexts"),
+      "@env": path.resolve(__dirname, "./src/env"),
+      "@features": path.resolve(__dirname, "./src/features"),
+      "@reducers": path.resolve(__dirname, "./src/features/reducers"),
+      "@fonts": path.resolve(__dirname, "./src/fonts"),
+      "@helpers": path.resolve(__dirname, "./src/helpers"),
+      "@hooks": path.resolve(__dirname, "./src/hooks"),
+      "@interfaces": path.resolve(__dirname, "./src/interfaces"),
+      "@layouts": path.resolve(__dirname, "./src/layouts"),
+      "@services": path.resolve(__dirname, "./src/services"),
+      "@less": path.resolve(__dirname, "./src/styles/less"),
+      "@utils": path.resolve(__dirname, "./src/utils")
     }
   },
   build: {
